# docs/api/api-contracts.yaml
openapi: 3.1.0
info:
  title: Enterprise AI Knowledge Assistant API
  description: |
    Production-grade RAG-powered knowledge assistant API.
    Enables semantic search and AI-powered Q&A over enterprise documents.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: platform@enterprise.com
servers:
  - url: https://api.knowledge.enterprise.com/v1
    description: Production
  - url: https://api-staging.knowledge.enterprise.com/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /documents:
    get:
      tags: [Documents]
      summary: List documents with pagination and filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, INDEXED, FAILED]
        - name: fileType
          in: query
          schema:
            type: string
            enum: [PDF, DOCX, TXT, MD, HTML]
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, title, fileSize]
            default: createdAt
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'

    post:
      tags: [Documents]
      summary: Upload a new document for ingestion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                category:
                  type: string
                metadata:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '202':
          description: Document accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '413':
          description: File too large
        '415':
          description: Unsupported file type
  /documents/bulk:
    post:
      tags: [Documents]
      summary: Upload multiple documents at once
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                category:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '202':
          description: Documents accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer
                  rejected:
                    type: integer
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentResponse'
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document details by ID
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
          description: Document details
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Documents]
      summary: Soft-delete a document and its embeddings
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204':
          description: Document deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/reindex:
    post:
      tags: [Documents]
      summary: Re-extract and re-embed a document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '202':
          description: Re-indexing started
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/download:
    get:
      tags: [Documents]
      summary: Get a pre-signed S3 download URL
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                    format: uri
                  expiresAt:
                    type: string
                    format: date-time
          description: Pre-signed URL

  # ─── CHAT ──────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSizeBytes:
          type: integer
        status:
          type: string
          enum: [PENDING, PROCESSING, INDEXED, FAILED]
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        chunkCount:
          type: integer
        pageCount:
          type: integer
        uploadedBy:
          type: string
        uploadedByName:
          type: string
        createdAt:
          type: string
          format: date-time
        indexedAt:
          type: string
          format: date-time
    DocumentPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
...